refactor(composer): extract the docked resize gesture
The pointer machinery for dragging the docked title bar is a self-contained
gesture: a stored height, a clamp, three pointer handlers and an outside-click
guard. It now lives in `useDockedResize`, fed the window mode, the body column
and the collapse action. Behaviour unchanged.
